Move useConfig composable to /composables/
